Fix #11: check npdId of cached license before using it
Before, when looking for a locally cached license against an operating config, we accepted any cached license for the config's app that had the right cert group.  But multiple configs can license the same app, so that would use the same cached license for configs from different packages.  Now we check the package ID in the license against the package ID of the config before we use the license for that config.